filter empty

+++ src/main/java/com/example/messagingstompwebsocket/model/dto/InitDto.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class InitDto {
   private MessageType messageType = MessageType.INIT;
   private List<List<String>> currentChain;
-  private List<Integer> totalPasses;
+  private List<Long> totalPasses;
   private List<String> players;
 }
